+++ src/mako/cli/output/components/Bell.php
@@ -5,14 +5,14 @@
  * @license   http://www.makoframework.com/license
  */
 
-namespace mako\cli\output\helpers;
+namespace mako\cli\output\components;
 
 use mako\cli\output\Output;
 
 use function str_repeat;
 
 /**
- * Bell helper.
+ * Bell component.
  */
 class Bell
 {