+++ src/mako/cli/output/components/OrderedList.php
@@ -5,7 +5,7 @@
  * @license   http://www.makoframework.com/license
  */
 
-namespace mako\cli\output\helpers;
+namespace mako\cli\output\components;
 
 use mako\cli\output\formatter\FormatterInterface;
 use mako\cli\output\Output;
@@ -16,7 +16,7 @@ use function str_repeat;
 use function strlen;
 
 /**
- * Ordered list helper.
+ * Ordered list component.
  */
 class OrderedList
 {