+++ src/mako/cli/output/components/Table.php
@@ -5,12 +5,12 @@
  * @license   http://www.makoframework.com/license
  */
 
-namespace mako\cli\output\helpers;
+namespace mako\cli\output\components;
 
 use mako\cli\exceptions\CliException;
+use mako\cli\output\components\table\Border;
+use mako\cli\output\components\traits\HelperTrait;
 use mako\cli\output\formatter\FormatterInterface;
-use mako\cli\output\helpers\table\Border;
-use mako\cli\output\helpers\traits\HelperTrait;
 use mako\cli\output\Output;
 
 use function array_values;
@@ -19,7 +19,7 @@ use function implode;
 use function str_repeat;
 
 /**
- * Table helper.
+ * Table component.
  */
 class Table
 {